better swagger + fix client issue
