+++ README.md
@@ -73,8 +73,8 @@ By default, the application is available at `http://localhost:7633`.
 Discover the core features to get started building and managing your data.
 Link to the official documentation: https://data.primals.net/en/documentation/
 
-- 🧠 [Core Concepts](./wiki/Concepts.md): Explore the fundamentals of data modeling
-- 🔌 [Custom API Endpoints](./wiki/Custom-Endpoints.md): Create dynamic HTTP routes directly from the backend
+- 🧠 [Core Concepts](https://github.com/anonympins/data-primals-engine/wiki/Concepts): Explore the fundamentals of data modeling
+- 🔌 [Custom API Endpoints](https://github.com/anonympins/data-primals-engine/wiki/Custom-Endpoints): Create dynamic HTTP routes directly from the backend
 - 🏗️ [Data Models](https://data.primals.net/en/documentation/create-models): Structure your information.
 - 🗃️ [Data Management](https://data.primals.net/en/documentation/manage-data): Create, read, update, and delete your entries.
 - 📊 [Views (Table, Kanban)](Views): Visualize your data in different ways. *(Page to be created)*
@@ -82,8 +82,6 @@ Link to the official documentation: https://data.primals.net/en/documentation/
 - 👥 [Users](https://data.primals.net/en/documentation/users): Manage access to the platform.
 - 🔐 [Roles and Permissions](https://data.primals.net/en/documentation/roles-permissions): Define who can see and do what.
 - ⚙️ [Automation with Workflows](https://data.primals.net/en/documentation/automation-workflows): Create automated processes.
-- 🎁 [Pack Gallery](Packs-gallery): Discover ready-to-use configurations. *(Page to be created)*
-- 🤖 [AI Assistance](AI-assistance): Leverage artificial intelligence. *(Page to be created)*
 
 ## 🔌 Integrate with your tools
 Connect the platform to your external applications and services via our API.
