+++ src/client.js
@@ -2,3 +2,5 @@
 // Ce fichier sert de point d'entrée pour les modules qui peuvent être utilisés en toute sécurité côté client (navigateur).
 
 export { Event } from './events.js';
+export { GameObject, Logger, BenchmarkTool } from './gameObject.js';
+export { Config } from './config.js';
